refactor: split common/functable.go — move type implementations to functable_types.go
Extract per-type method implementations from functable.go into a dedicated
functable_types.go file to reduce file size and improve navigability.
